Fix room names to not include a space.
Adding the space is unreliable without making exceptions for certain
building abbreviations.

Some examples:
- M31006 would be split to "M 31006" instead of "M3 1006". We could
solve this one by ensuring all building names are two or more
characters (and allow the second character to be a number).
- EV34412 would be split into "EV 34412" instead of "EV3 4412", which
isn't possible to split without making an exception for the EV[1-3]
buildings.

I don't really feel it's worth making special cases at this point, so
we'll just do without the space for now.